Update services and modify configs for database integration
Refactored workflows and DTOs to be used as a library and relocated to libs folder. Introduced TypeORM integration in workflows service and virtual facility for database operations. Several services and DTOs have been updated to fit with these changes. Postgres configurations were added in the app.module.ts. A Dockerfile was also added for containerization.